+++ app/Orm/UserExpensesMapper.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace Nexendrie\Orm;
+
+/**
+ * @author Jakub Konečný
+ */
+final class UserExpensesMapper extends \Nextras\Orm\Mapper\Mapper {
+}
+?>